automation/requirements.txt
Appium_Python_Client==4.0.0
nltk==3.8.1
pandas==2.2.2
selenium==4.20.0
tqdm==4.65.0
XlsxWriter==3.2.0
openpyxl==3.1.2
matplotlib==3.8.4
